* The portlet-model-hints.xml file now gets generated as well. * Multi-word command line arguments can be given in quotes

+++ db2servicexml/src/main/java/fi/jmvh/liferay/db2servicexml/db/model/SQLDatatypeEnum.java
@@ -10,7 +10,7 @@ package fi.jmvh.liferay.db2servicexml.db.model;
  */
 public enum SQLDatatypeEnum {
     varchar("String"),
-    numeric("int");
+    numeric("long");
     
     private String javaType;
